YAY-13 refactor: remove logic from component into useCroppingModal
