+++ src/entities/posts/ui/modals/addPhotoModal/AddPhotoModal.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { PostOutlineIcon } from '@/shared/icons'
 import { Button } from '@/shared/components/Button'
 import { ChangeEvent, useEffect, useRef } from 'react'
