+++ src/entities/posts/ui/modals/croppingModal/CroppingModal.tsx
@@ -1,7 +1,5 @@
-'use client'
 import { PostModal } from '@/shared/components/PostModal'
 import Image, { StaticImageData } from 'next/image'
-import { ChangeEvent, ReactNode, useCallback, useEffect, useReducer, useRef, useState } from 'react'
 import {
    CrossIcon,
    ExpandOutline,
@@ -11,14 +9,11 @@ import {
 } from '@/shared/icons'
 import { cn } from '@/shared/lib'
 import { RangeControl } from '@/shared/components/RangeControl'
-import type { Area } from 'react-easy-crop'
 import Cropper from 'react-easy-crop'
-import { croppingReducer, initialState } from './croppingReducer'
-import { loadImageMeta } from '@/shared/utils'
 import { Slider } from '@/shared/components/Slider'
-import { getCroppedImg } from '@/shared/utils/images/cropImage'
-import { ACCEPTED_IMAGE_TYPES, MAX_FILE_SIZE } from '@/shared/schema'
-import { alert } from '@/shared/components/Alert'
+import { ACCEPTED_IMAGE_TYPES } from '@/shared/schema'
+import { useCroppingModal } from '@/entities/posts/ui/modals/croppingModal/useCroppingModal'
+import { AspectOption } from '@/entities/posts/ui/modals/croppingModal/AspectOption'
 
 type Props = {
    isOpen: boolean
@@ -39,183 +34,42 @@ export const CroppingModal = ({
    onOpenChange,
    setImages,
 }: Props) => {
-   const [state, dispatch] = useReducer(croppingReducer, initialState)
-   const { showZoomScale, showAspectRatio, showImageGallery, imageStates } = state
-   const [isProcessing, setIsProcessing] = useState(false)
-
-   // Получаем состояние для текущего изображения
-   const currentImageState = imageStates[currentIndex] || {
-      crop: { x: 0, y: 0 },
-      zoomScale: [20],
-      aspect: undefined,
-      naturalAspect: undefined,
-      croppedAreaPixels: null,
-   }
-
-   const { crop, zoomScale, aspect, naturalAspect } = currentImageState
-
-   const fileInputRef = useRef<HTMLInputElement | null>(null)
-
-   const handleAddImageClick = () => {
-      fileInputRef.current?.click()
-   }
-
-   const handleFilesSelected = async (e: ChangeEvent<HTMLInputElement>) => {
-      if (!e.target.files?.length) return
-
-      const file = e.target.files[0]
-
-      if (file.size > MAX_FILE_SIZE) {
-         alert.error('File size exceeds 5MB. Please choose a smaller file.')
-         return
-      }
-
-      const src = URL.createObjectURL(file)
-
-      const { width, height } = await loadImageMeta(src)
-
-      setImages([...images, { src, width, height }])
-   }
-
-   const deleteImage = (idxToDelete: number) => {
-      const newImages = images.filter((_, idx) => idx !== idxToDelete)
-      setImages(newImages)
-
-      if (currentIndex >= newImages.length) {
-         setCurrentIndex(Math.max(newImages.length - 1, 0))
-      }
-   }
-
-   const handleThumbnailClick = (idx: number) => {
-      setCurrentIndex(idx)
-   }
-
-   const image = images[currentIndex]
-
-   const zoom = 1 + (zoomScale[0] / 100) * 2
-
-   // Инициализируем состояние для изображения
-   useEffect(() => {
-      if (!image) return
-
-      const naturalAspectValue = image.width / image.height
-
-      if (!imageStates[currentIndex]) {
-         dispatch({
-            type: 'INIT_IMAGE_STATE',
-            payload: {
-               index: currentIndex,
-               naturalAspect: naturalAspectValue,
-            },
-         })
-      }
-   }, [image, currentIndex, imageStates])
-
-   // Сохраняем финальные данные об обрезке
-   const onCropComplete = useCallback(
-      (_: Area, croppedAreaPixels: Area) => {
-         dispatch({
-            type: 'SET_CROPPED_AREA',
-            payload: {
-               index: currentIndex,
-               croppedAreaPixels,
-            },
-         })
-      },
-      [currentIndex]
-   )
-
-   const handleCropChange = useCallback(
-      (crop: { x: number; y: number }) => {
-         dispatch({
-            type: 'SET_CROP',
-            payload: {
-               index: currentIndex,
-               crop,
-            },
-         })
-      },
-      [currentIndex]
-   )
-
-   const handleChangeRange = (newZoomScale: number[]) => {
-      dispatch({
-         type: 'SET_ZOOM_SCALE',
-         payload: {
-            index: currentIndex,
-            zoomScale: newZoomScale,
-         },
-      })
-   }
-
-   const handleZoomChange = (newZoom: number) => {
-      const newZoomScaleValue = [((newZoom - 1) / 2) * 100]
-      dispatch({
-         type: 'SET_ZOOM_SCALE',
-         payload: {
-            index: currentIndex,
-            zoomScale: newZoomScaleValue,
-         },
-      })
-   }
-
-   const toggleAspectRatio = () => {
-      dispatch({ type: 'TOGGLE_ASPECT_RATIO' })
-   }
-
-   const toggleZoomScale = () => dispatch({ type: 'TOGGLE_ZOOM_SCALE' })
-
-   const toggleGallery = () => dispatch({ type: 'TOGGLE_IMAGE_GALLERY' })
-
-   const handleSliderNavigation = (newIndex: number) => {
-      setCurrentIndex(newIndex)
-   }
-
-   // Функция для обрезки всех изображений и сохранения их в состоянии
-   const processAndSaveCroppedImages = async (): Promise<StaticImageData[]> => {
-      return await Promise.all(
-         images.map(async (image, index) => {
-            const imageState = imageStates[index]
-
-            // Если есть данные об обрезке - создаем обрезанное изображение
-            if (imageState?.croppedAreaPixels) {
-               try {
-                  const croppedSrc = await getCroppedImg(image.src, imageState.croppedAreaPixels)
-                  return {
-                     src: croppedSrc,
-                     width: imageState.croppedAreaPixels.width,
-                     height: imageState.croppedAreaPixels.height,
-                  }
-               } catch (error) {
-                  console.error('Error cropping image:', error)
-                  // В случае ошибки возвращаем оригинальное изображение
-                  return image
-               }
-            } else {
-               // Если обрезки не было - возвращаем оригинал
-               return image
-            }
-         })
-      )
-   }
-
-   // Обработчик для кнопки Next - обрезаем изображения и сохраняем
-   const handleNext = async () => {
-      setIsProcessing(true)
-      try {
-         const croppedImages = await processAndSaveCroppedImages()
-         // Сохраняем обрезанные изображения в состоянии
-         setImages(croppedImages)
-         // Вызываем оригинальный onNext
-         onNext()
-      } catch (error) {
-         console.error('Error processing images:', error)
-         // В случае ошибки просто переходим дальше с оригинальными изображениями
-         onNext()
-      } finally {
-         setIsProcessing(false)
-      }
-   }
+   const {
+      showZoomScale,
+      showAspectRatio,
+      showImageGallery,
+      isProcessing,
+      crop,
+      zoom,
+      aspect,
+      naturalAspect,
+      image,
+      fileInputRef,
+      handleAddImageClick,
+      handleSet16_9Aspect,
+      handleSetOriginalAspect,
+      handleSet1_1Aspect,
+      handleSet4_5Aspect,
+      handleFilesSelected,
+      deleteImage,
+      handleThumbnailClick,
+      onCropComplete,
+      handleCropChange,
+      handleChangeRange,
+      handleZoomChange,
+      toggleAspectRatio,
+      toggleZoomScale,
+      toggleGallery,
+      handleSliderNavigation,
+      handleNext,
+      zoomScale,
+   } = useCroppingModal({
+      images,
+      currentIndex,
+      setCurrentIndex,
+      setImages,
+      onNext,
+   })
 
    const baseInteractiveButtonStyle =
       'cursor-pointer w-9 h-9 rounded-xs bg-dark-500 absolute bottom-[11px] flex items-center justify-center opacity-80'
@@ -309,15 +163,7 @@ export const CroppingModal = ({
                   label="Original"
                   value={naturalAspect}
                   aspect={aspect}
-                  onClick={() =>
-                     dispatch({
-                        type: 'SET_ASPECT',
-                        payload: {
-                           index: currentIndex,
-                           aspect: naturalAspect,
-                        },
-                     })
-                  }
+                  onClick={handleSetOriginalAspect}
                   Icon={<ImageOutline />}
                />
 
@@ -325,15 +171,7 @@ export const CroppingModal = ({
                   label="1:1"
                   value={1}
                   aspect={aspect}
-                  onClick={() =>
-                     dispatch({
-                        type: 'SET_ASPECT',
-                        payload: {
-                           index: currentIndex,
-                           aspect: 1,
-                        },
-                     })
-                  }
+                  onClick={handleSet1_1Aspect}
                   boxClass="h-[18px] w-[18px]"
                />
 
@@ -341,15 +179,7 @@ export const CroppingModal = ({
                   label="4:5"
                   value={4 / 5}
                   aspect={aspect}
-                  onClick={() =>
-                     dispatch({
-                        type: 'SET_ASPECT',
-                        payload: {
-                           index: currentIndex,
-                           aspect: 4 / 5,
-                        },
-                     })
-                  }
+                  onClick={handleSet4_5Aspect}
                   boxClass="h-[26px] w-[18px]"
                />
 
@@ -357,15 +187,7 @@ export const CroppingModal = ({
                   label="16:9"
                   value={16 / 9}
                   aspect={aspect}
-                  onClick={() =>
-                     dispatch({
-                        type: 'SET_ASPECT',
-                        payload: {
-                           index: currentIndex,
-                           aspect: 16 / 9,
-                        },
-                     })
-                  }
+                  onClick={handleSet16_9Aspect}
                   boxClass="h-[20px] w-[26px]"
                />
             </div>
@@ -447,47 +269,5 @@ export const CroppingModal = ({
             </div>
          )}
       </PostModal>
-   )
-}
-
-function AspectOption({
-   label,
-   value,
-   aspect,
-   Icon,
-   boxClass,
-   onClick,
-}: {
-   label: string
-   value: number | undefined
-   aspect: number | undefined
-   Icon?: ReactNode
-   boxClass?: string
-   onClick: () => void
-}) {
-   const isActive = value === aspect
-
-   return (
-      <button
-         onClick={onClick}
-         className={cn(
-            'flex h-[38px] w-full cursor-pointer items-center justify-between px-3 py-2.5',
-            isActive ? 'text-light-100' : 'text-light-900'
-         )}
-      >
-         {label}
-
-         {Icon ? (
-            Icon
-         ) : (
-            <div
-               className={cn(
-                  'relative right-[3px] inline-block rounded-xs border-2',
-                  boxClass,
-                  isActive ? 'border-light-100' : 'border-light-900'
-               )}
-            />
-         )}
-      </button>
    )
 }
