+++ src/shared/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './handleError'
 export * from './typeguards/isErrorWithMessage'
 export * from './images/loadImageMeta'
-export * from '../hook/useCreateQueryString'
+export * from './images/cropImage'
